pyproject.toml
[project]
name = "WeVolunteer"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "coverage>=7.11.0",
    "datastar>=0.0.8",
    "datastar-py>=0.5.0",
    "dj-database-url>=3.0.1",
    "django>=5.2.3",
    "django-allauth[socialaccount]>=65.9.0",
    "django-bootstrap-datepicker-plus>=5.0.5",
    "django-google-fonts>=0.0.3",
    "django-template-partials>=24.4",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "rules>=3.5",
    "whitenoise>=6.9.0",
]
